PRE-ALPHA_1.0.3_HOTFIX (Added much needed disclaimers to the data\\apps\\ folder, read release notes for more info.)
Added OBVIOUS DISCLAIMERS to the player's apps folder.
Some .sh files contain malicous code that is intended to be run in-game,
and NOT on your actual PC.
Thank you @cpt-ado for bringing this concern to my attention.
Do note: This is a temporary measure. I will be moving the contents of
ALL files win the data\\apps\\ folder into a single .json file for easy
reading, and to prevent accidental running of these .sh files. (These
files were meant to be fictional, but they 'could' actually be run, and
cause dammage to a person's system.)